Begin to split up ComponentProcessorTest into manageable chunks
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=195122058
